Front-end da aplicação

+++ backend/src/controllers/IncidentController.js
@@ -45,9 +45,13 @@ module.exports = {
             .where('id', id)
             .select('ong_id')
             .first();
+        
+        if(!incident){
+            return response.status(400).json({ error: 'Incident inexisting.'});
+        }
 
         if (incident.ong_id != ong_id){
-            return response.status(401).json({ error: 'Operation not permitted'});
+            return response.status(401).json({ error: 'Operation not permitted.'});
         }
 
         await connection('incidents').where('id', id).delete();
